requirements.txt
flask==3.0.0
supabase
python-dotenv
